+++ src/unfold/admin.py
@@ -1,7 +1,9 @@
 from functools import update_wrapper
-from typing import Any
+from typing import Any, TypedDict
 
 from django import forms
+from django.conf import settings
+from django.contrib import messages
 from django.contrib.admin import ModelAdmin as BaseModelAdmin
 from django.contrib.admin import StackedInline as BaseStackedInline
 from django.contrib.admin import TabularInline as BaseTabularInline
@@ -14,15 +16,14 @@ from django.contrib.contenttypes.admin import (
     GenericTabularInline as BaseGenericTabularInline,
 )
 from django.db.models import BLANK_CHOICE_DASH, Model
-from django.http import HttpRequest
-from django.template.response import TemplateResponse
+from django.http import HttpRequest, HttpResponse
 from django.urls import URLPattern, path
-from django.utils.safestring import mark_safe
+from django.utils.html import format_html
+from django.utils.safestring import SafeString, mark_safe
 from django.utils.translation import gettext_lazy as _
 from django.views import View
 
 from unfold.checks import UnfoldModelAdminChecks
-from unfold.datasets import BaseDataset
 from unfold.forms import (
     ActionForm,
     PaginationGenericInlineFormSet,
@@ -30,12 +31,12 @@ from unfold.forms import (
 )
 from unfold.mixins import (
     ActionModelAdminMixin,
-    BaseModelAdminMixin,
     DatasetModelAdminMixin,
+    FormFieldModelAdminMixin,
     NestedInlinesModelAdminMixin,
 )
 from unfold.overrides import FORMFIELD_OVERRIDES_INLINE
-from unfold.typing import FieldsetsType
+from unfold.utils import get_setting_value
 from unfold.views import ChangeList
 from unfold.widgets import UnfoldBooleanWidget
 
@@ -48,8 +49,13 @@ checkbox = UnfoldBooleanWidget(
 )
 
 
+class ListFilterOptionsItem(TypedDict):
+    label: str | None
+    horizontal: bool | None
+
+
 class ModelAdmin(
-    BaseModelAdminMixin,
+    FormFieldModelAdminMixin,
     ActionModelAdminMixin,
     DatasetModelAdminMixin,
     NestedInlinesModelAdminMixin,
@@ -63,6 +69,7 @@ class ModelAdmin(
     list_horizontal_scrollbar_top = False
     list_filter_submit = False
     list_filter_sheet = True
+    list_filter_options: dict[str, ListFilterOptionsItem] = {}
     list_fullwidth = False
     list_disable_select_all = False
     list_before_template = None
@@ -71,7 +78,6 @@ class ModelAdmin(
     change_form_after_template = None
     change_form_outer_before_template = None
     change_form_outer_after_template = None
-    change_form_datasets = ()
     compressed_fields = True
     show_add_link = True
     readonly_preprocess_fields = {}
@@ -102,7 +108,7 @@ class ModelAdmin(
 
     def changelist_view(
         self, request: HttpRequest, extra_context: dict[str, str] | None = None
-    ) -> TemplateResponse:
+    ) -> HttpResponse:
         self.request = request
 
         if self.ordering_field and self.ordering_field not in self.list_editable:
@@ -111,6 +117,39 @@ class ModelAdmin(
             self.list_editable = list_editable
 
         return super().changelist_view(request, extra_context)
+
+    def changeform_view(
+        self,
+        request: HttpRequest,
+        object_id: str | None = None,
+        form_url: str = "",
+        extra_context: dict[str, Any] | None = None,
+    ) -> Any:
+        from unfold.forms import AdminForm, Fieldline
+
+        helpers.AdminForm = AdminForm  # ty:ignore
+        helpers.Fieldline = Fieldline  # ty:ignore
+
+        response = super().changeform_view(request, object_id, form_url, extra_context)
+
+        if (
+            request.method == "GET"
+            and settings.DEBUG
+            and get_setting_value("SHOW_UI_WARNINGS", request) is True
+        ):
+            for missing_field in sorted(set(self._autocomplete_fields_missing)):
+                self.message_user(
+                    request,
+                    format_html(
+                        _(
+                            'Field <strong class="font-semibold">{field_name}</strong> is not an autocomplete field. Please add it to the `autocomplete_fields` list.'
+                        ),  # ty:ignore[invalid-argument-type]
+                        field_name=missing_field,
+                    ),
+                    messages.WARNING,
+                )
+
+        return response
 
     def get_list_display(self, request: HttpRequest) -> list | tuple:
         list_display = super().get_list_display(request)
@@ -125,7 +164,7 @@ class ModelAdmin(
 
     def get_fieldsets(
         self, request: HttpRequest, obj: Model | None = None
-    ) -> FieldsetsType:
+    ) -> list | tuple:
         if not obj and self.add_fieldsets:
             return self.add_fieldsets
         return super().get_fieldsets(request, obj)
@@ -206,10 +245,10 @@ class ModelAdmin(
         return super().get_action_choices(request, default_choices)
 
     @display(description=mark_safe(checkbox.render("action_toggle_all", 1)))
-    def action_checkbox(self, obj: Model) -> str:
+    def action_checkbox(self, obj: Model) -> SafeString:
         return checkbox.render(helpers.ACTION_CHECKBOX_NAME, str(obj.pk))
 
-    def get_changelist(self, request: HttpRequest, **kwargs: Any) -> ChangeList:
+    def get_changelist(self, request: HttpRequest, **kwargs: Any) -> type[ChangeList]:
         return ChangeList
 
     def get_formset_kwargs(
@@ -232,9 +271,6 @@ class ModelAdmin(
 
         return formset_kwargs
 
-    def get_changeform_datasets(self, request: HttpRequest) -> list[type[BaseDataset]]:
-        return self.change_form_datasets
-
 
 class BaseInlineMixin:
     formfield_overrides = FORMFIELD_OVERRIDES_INLINE
@@ -248,21 +284,21 @@ class BaseInlineMixin:
     tab = False
 
 
-class TabularInline(BaseInlineMixin, BaseModelAdminMixin, BaseTabularInline):
+class TabularInline(BaseInlineMixin, FormFieldModelAdminMixin, BaseTabularInline):
     formset = PaginationInlineFormSet
 
 
-class StackedInline(BaseInlineMixin, BaseModelAdminMixin, BaseStackedInline):
+class StackedInline(BaseInlineMixin, FormFieldModelAdminMixin, BaseStackedInline):
     formset = PaginationInlineFormSet
 
 
 class GenericStackedInline(
-    BaseInlineMixin, BaseModelAdminMixin, BaseGenericStackedInline
+    BaseInlineMixin, FormFieldModelAdminMixin, BaseGenericStackedInline
 ):
     formset = PaginationGenericInlineFormSet
 
 
 class GenericTabularInline(
-    BaseInlineMixin, BaseModelAdminMixin, BaseGenericTabularInline
+    BaseInlineMixin, FormFieldModelAdminMixin, BaseGenericTabularInline
 ):
     formset = PaginationGenericInlineFormSet